+++ tests/test_security.py
@@ -1,5 +1,6 @@
 import os
 import posixpath
+import sys
 
 import pytest
 
@@ -8,25 +9,37 @@ from werkzeug.security import generate_password_hash
 from werkzeug.security import safe_join
 
 
-def test_password_hashing():
-    hash0 = generate_password_hash("default")
-    assert check_password_hash(hash0, "default")
-    assert hash0.startswith("pbkdf2:sha256:600000$")
+def test_default_password_method():
+    value = generate_password_hash("secret")
+    assert value.startswith("pbkdf2:")
 
-    hash1 = generate_password_hash("default", "sha1")
-    hash2 = generate_password_hash("default", method="sha1")
+
+@pytest.mark.xfail(
+    sys.implementation.name == "pypy", reason="scrypt unavailable on pypy"
+)
+def test_scrypt():
+    value = generate_password_hash("secret", method="scrypt")
+    assert check_password_hash(value, "secret")
+    assert value.startswith("scrypt:32768:8:1$")
+
+
+def test_pbkdf2():
+    value = generate_password_hash("secret", method="pbkdf2")
+    assert check_password_hash(value, "secret")
+    assert value.startswith("pbkdf2:sha256:600000$")
+
+
+def test_salted_hashes():
+    hash1 = generate_password_hash("secret")
+    hash2 = generate_password_hash("secret")
     assert hash1 != hash2
-    assert check_password_hash(hash1, "default")
-    assert check_password_hash(hash2, "default")
-    assert hash1.startswith("sha1$")
-    assert hash2.startswith("sha1$")
+    assert check_password_hash(hash1, "secret")
+    assert check_password_hash(hash2, "secret")
 
+
+def test_require_salt():
     with pytest.raises(ValueError):
-        generate_password_hash("default", "sha1", salt_length=0)
-
-    fakehash = generate_password_hash("default", method="plain")
-    assert fakehash == "plain$$default"
-    assert check_password_hash(fakehash, "default")
+        generate_password_hash("secret", salt_length=0)
 
 
 def test_safe_join():